refactor(case): remove prologue back-fill and its Offer(CaseManagerRole) coupling
ADR-0041 makes the CaseActor the authoritative creator of the case, so the
vendor-authored prologue back-fill is no longer needed.

- Delete WritePrologueLedgerEntriesNode (nodes/prologue.py) and its test
- Relocate the canonical payloadSnapshot builders to
  case/ledger_snapshots.py with public names; case_proposal_received_tree
  uses them to commit the four init entries natively (CM-22-003)
- Remove the prologue node from offer_case_manager_role_received_tree; keep
  the accept/reject handshake functional as a standalone spec-mandated
  operation (DEMOMA-08-002/003/006-009) driven by its manual trigger, and
  document that decision in the module docstring
- Add ("Add", "CaseStatus") to _CASE_AUTHORED_SIGNATURES per CLP-12-001 so
  CaseActor-authored add_case_status_to_case entries validate (#1767)
- Extract canonical-entry validation from chain.py into
  sync/nodes/canonical_entry.py to stay under the BTND-07-004 500-line cap
- Update specs/notes pointers to the new module locations

Closes #1777
Closes #1767
