+++ vultron/core/behaviors/sync/nodes/chain.py
@@ -23,6 +23,9 @@ import py_trees
 from py_trees.common import Status
 
 from vultron.core.behaviors.helpers import DataLayerAction
+from vultron.core.behaviors.sync.nodes.canonical_entry import (
+    _validate_canonical_entry,
+)
 from vultron.core.models._helpers import _now_utc
 from vultron.core.models.case_ledger import HashChainLedgerRecord
 from vultron.core.models.case_ledger_entry import CaseLedgerEntry
@@ -30,59 +33,10 @@ from vultron.core.models.case_ledger_entry import VultronCaseLedgerEntry
 from vultron.core.models.replication_state import VultronReplicationState
 from vultron.core.sync_helpers import _find_equivalent_recorded_entry
 from vultron.core.sync_helpers import _reconstruct_tail_hash
-from vultron.errors import VultronCanonicalEntryError
 from vultron.errors import VultronError
 from vultron.errors import VultronValidationError
 
 logger = logging.getLogger(__name__)
-
-_CANONICAL_PAYLOAD_SIGNATURES: tuple[tuple[str, str], ...] = (
-    ("Create", "VulnerabilityCase"),
-    ("Offer", "VulnerabilityReport"),
-    ("Offer", "VulnerabilityCase"),
-    ("Accept", "Offer"),  # validate_report (RV) — object_ is the Offer
-    ("TentativeReject", "Offer"),  # invalidate_report (RI)
-    ("Reject", "Offer"),  # close_report (RC)
-    ("Read", "Offer"),  # ack_report (RK, ADR-0021)
-    ("Add", "Note"),
-    ("Add", "VulnerabilityReport"),  # add_report_to_case
-    ("Add", "CaseStatus"),  # add_case_status_to_case
-    ("Add", "ParticipantStatus"),
-    ("Add", "EmbargoEvent"),
-    ("Remove", "EmbargoEvent"),
-    ("Offer", "EmbargoEvent"),
-    ("Invite", "EmbargoEvent"),
-    ("Accept", "EmbargoEvent"),
-    ("Reject", "EmbargoEvent"),
-    ("Join", "VulnerabilityCase"),
-    ("Ignore", "VulnerabilityCase"),
-    ("Leave", "VulnerabilityCase"),
-    ("Invite", "VulnerabilityCase"),
-    ("Accept", "Invite"),
-    ("Reject", "Invite"),
-    ("Announce", "VulnerabilityCase"),
-    ("Offer", "CaseParticipant"),
-    ("Add", "CaseParticipant"),
-)
-_CASE_AUTHORED_SIGNATURES: frozenset[tuple[str, str]] = frozenset(
-    {
-        ("Announce", "VulnerabilityCase"),
-        ("Add", "EmbargoEvent"),
-        ("Remove", "EmbargoEvent"),
-        ("Invite", "EmbargoEvent"),
-        ("Offer", "CaseParticipant"),
-        ("Invite", "VulnerabilityCase"),
-        ("Offer", "VulnerabilityCase"),
-        ("Leave", "VulnerabilityCase"),
-        ("Accept", "Offer"),
-        ("Reject", "Offer"),
-        ("Add", "CaseParticipant"),
-        ("Create", "VulnerabilityCase"), ("Add", "VulnerabilityReport"), ("Add", "ParticipantStatus"),  # noqa: E501  # fmt: skip
-    }
-)
-_INLINE_OBJECT_KEYS: frozenset[str] = frozenset(
-    {"object", "object_", "target"}
-)
 
 
 def _require_log_entry(
@@ -129,120 +83,6 @@ def _to_persistable_entry(
         reason_code=chain_entry.reason_code,
         reason_detail=chain_entry.reason_detail,
     )
-
-
-def _snapshot_type(snapshot: dict[str, Any]) -> str | None:
-    activity_type = snapshot.get("type") or snapshot.get("type_")
-    return (
-        activity_type
-        if isinstance(activity_type, str) and activity_type
-        else None
-    )
-
-
-_ACTOR_TYPES: frozenset[str] = frozenset(
-    {"Actor", "Application", "Group", "Organization", "Person", "Service"}
-) | {"CoreActor"}
-
-
-def _snapshot_object_type(snapshot: dict[str, Any]) -> str | None:
-    # Invite(Actor, target=Case): object_ is the actor; use target.type so the
-    # signature resolves to ('Invite','VulnerabilityCase') not ('Invite','Org').
-    obj = snapshot.get("object") or snapshot.get("object_")
-    if not isinstance(obj, dict):
-        return None
-    object_type = obj.get("type") or obj.get("type_")
-    if not isinstance(object_type, str) or not object_type:
-        return None
-    if object_type in _ACTOR_TYPES:
-        target = snapshot.get("target")
-        if isinstance(target, dict):
-            target_type = target.get("type") or target.get("type_")
-            if isinstance(target_type, str) and target_type:
-                return target_type
-    return object_type
-
-
-def _bare_inline_object_path(
-    value: Any, path: str = "payloadSnapshot"
-) -> str | None:
-    if isinstance(value, dict):
-        for key, child in value.items():
-            child_path = f"{path}.{key}"
-            if key in _INLINE_OBJECT_KEYS and isinstance(child, str):
-                return child_path
-            nested_path = _bare_inline_object_path(child, child_path)
-            if nested_path is not None:
-                return nested_path
-        return None
-    if isinstance(value, list):
-        for index, item in enumerate(value):
-            nested_path = _bare_inline_object_path(item, f"{path}[{index}]")
-            if nested_path is not None:
-                return nested_path
-    return None
-
-
-def _validate_canonical_entry(
-    *,
-    case_id: str,
-    actor_id: str | None,
-    case_actor_id: str | None = None,
-    disposition: str,
-    payload_snapshot: dict[str, Any],
-    event_type: str,
-) -> None:
-    # Runs before idempotency check so malformed entries never reach the
-    # equivalence lookup (CLP-07). Relaxed for non-recorded dispositions.
-    if disposition != "recorded":
-        return
-    if not payload_snapshot:
-        raise VultronCanonicalEntryError(
-            f"{event_type}: recorded canonical entries require a non-empty "
-            "payloadSnapshot"
-        )
-
-    snapshot_actor = payload_snapshot.get("actor")
-    if not isinstance(snapshot_actor, str) or not snapshot_actor:
-        raise VultronCanonicalEntryError(
-            f"{event_type}: payloadSnapshot.actor must be a non-empty URI"
-        )
-
-    activity_type = _snapshot_type(payload_snapshot)
-    object_type = _snapshot_object_type(payload_snapshot)
-    signature = (activity_type or "", object_type or "")
-
-    bare_reference_path = _bare_inline_object_path(payload_snapshot)
-    if bare_reference_path is not None:
-        raise VultronCanonicalEntryError(
-            f"{event_type}: {bare_reference_path} must be an inline object, "
-            "not a bare ID string"
-        )
-
-    if signature not in _CANONICAL_PAYLOAD_SIGNATURES:
-        raise VultronCanonicalEntryError(
-            f"{event_type}: payloadSnapshot type/object pair {signature!r} "
-            "is not canonical"
-        )
-
-    # CLP-07-003: only CaseActor-authored activities may have the CaseActor as
-    # snapshot actor; all participant-originated activities must have a
-    # participant (non-CaseActor) actor.
-    if (
-        case_actor_id
-        and snapshot_actor == case_actor_id
-        and signature not in _CASE_AUTHORED_SIGNATURES
-    ):
-        raise VultronCanonicalEntryError(
-            f"{event_type}: payloadSnapshot.actor must not be the CaseActor"
-            f" for non-case-authored entries (signature={signature!r})"
-        )
-
-    context = payload_snapshot.get("context")
-    if context != case_id:
-        raise VultronCanonicalEntryError(
-            f"{event_type}: payloadSnapshot.context must equal the case URI"
-        )
 
 
 class ReconstructChainTailNode(DataLayerAction):
